requirements.txt
boto3==1.39.4
botocore==1.39.4
jmespath==1.0.1
pillow==11.3.0
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
urllib3==2.5.0
